+++ coursesModule/src/main/java/com/example/coursesmodule/model/Approfundation.java
@@ -1,20 +1,61 @@
 package com.example.coursesmodule.model;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.validation.constraints.NotNull;
+
+import java.util.ArrayList;
 import java.util.List;
 
-public interface Approfundation {
+public class Approfundation {
+
+    private final String type;
+    @NotNull
+    private int id;
+    private List<Resource> resources;
+    @NotNull
+    private int numberWeeks;
+
+    public Approfundation(@JsonProperty("id") int id, @JsonProperty("numberWeeks") int numberWeeks) {
+        this.resources = new ArrayList<>();
+        this.numberWeeks = numberWeeks;
+        this.id = id;
+        this.type = "Laboratory";
+    }
 
     //setters
-    public void setResources(List<Resource> resources);
-    public void setNumberWeeks(int numberWeeks);
-    public void setId(int id);
+    public void setResources(List<Resource> resources) {
+        this.resources = resources;
+    }
+    public void setNumberWeeks(int numberWeeks) {
+        this.numberWeeks = numberWeeks;
+    }
+
+
+    public void setId(int id) {
+        this.id = id;
+    }
 
     //getters
-    public String getType();
-    public List<Resource> getResources();
-    public int getNumberWeeks();
-    public int getId();
+    public String getType() {
+        return type;
+    }
+    public List<Resource> getResources() {
+        return resources;
+    }
+    public int getNumberWeeks() {
+        return numberWeeks;
+    }
+
+    public int getId() {
+        return id;
+    }
 
     //additional methods
-    public void addResource(Resource resource);
+    public void addResource(Resource resource) {
+        this.resources.add(resource);
+    }
+    public void removeResource(Resource resource) {
+        this.resources.remove(resource);
+    }
+
 }
