--- coursesModule/src/main/java/com/example/coursesmodule/model/Seminar.java
@@ -1,63 +1,0 @@
-package com.example.coursesmodule.model;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import jakarta.validation.constraints.NotNull;
-
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class Seminar implements Approfundation {
-
-    private final String type;
-    @NotNull
-    private int id;
-    private List<Resource> resources;
-
-    @NotNull
-    private int numberWeeks;
-
-    public Seminar(@JsonProperty("id") int id,@JsonProperty("numberWeeks") int numberWeeks) {
-        this.resources = new ArrayList<>();
-        this.numberWeeks = numberWeeks;
-        this.id = id;
-        this.type = "Seminar";
-    }
-
-    //setters
-    public void setResources(List<Resource> resources) {
-        this.resources = resources;
-    }
-    public void setNumberWeeks(int numberWeeks) {
-        this.numberWeeks = numberWeeks;
-    }
-
-    @Override
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    //getters
-    public String getType() {
-        return type;
-    }
-    public List<Resource> getResources() {
-        return resources;
-    }
-    public int getNumberWeeks() {
-        return numberWeeks;
-    }
-
-    @Override
-    public int getId() {
-        return id;
-    }
-
-    //additional methods
-    public void addResource(Resource resource) {
-        this.resources.add(resource);
-    }
-    public void removeResource(Resource resource) {
-        this.resources.remove(resource);
-    }
-}
